+++ server.mjs
@@ -5,6 +5,7 @@ import { Server as ServerIO } from "socket.io";
 
 
 const dev = process.env.NODE_ENV !== 'production';
+// const hostname = 'liltrees.onrender.com';
 const hostname = 'localhost';
 console.log(process.env.NODE_ENV)
 const port = process.env.PORT || 3000;
@@ -57,20 +58,20 @@ app.prepare().then(() => {
 
 
     socket.on("channel-input",function(arg1_channelId,arg2_message_item,arg3_type_channel) {
-      console.log("data from channel input",arg1_channelId,arg2_message_item.content,arg3_type_channel);
+      // console.log("data from channel input",arg1_channelId,arg2_message_item.content,arg3_type_channel);
       // socket.broadcast.emit(arg1,arg2);
       io.emit(arg1_channelId,arg2_message_item,arg3_type_channel)
       })
   
-      socket.on("channel-update",function(arg1_channelId,arg3_type_channel) {
-        console.log("data from channel update",arg1_channelId,arg3_type_channel);
-        // socket.broadcast.emit(arg1,arg2);
-        io.emit(arg1_channelId,arg3_type_channel)
-        })
+    socket.on("channel-update",function(arg1_channelId,arg3_type_channel) {
+      // console.log("data from channel update",arg1_channelId,arg3_type_channel);
+      // socket.broadcast.emit(arg1,arg2);
+      io.emit(arg1_channelId,arg3_type_channel)
+      })
   
     socket.onAny((event, ...args) => {
       console.log(`got ${event}`);
-      console.log(`data from server.mjs is ${args}`);
+      // console.log(`data from server.mjs is ${args}`);
     });
   });
 })